backend/requirements.txt
pandas==1.3.1
numpy==1.21.1
psycopg2-binary==2.9.1
sqlalchemy==1.4.22
celery==5.1.2
scikit-learn==0.24.2
flask==2.0.1
Flask-Cors==3.0.10
mlflow==1.19.0
pika==1.2.0
pandas-profiling==3.1.0
xgboost==1.4.2
torch==1.8.1
imbalanced-learn==0.8.0
mlprep-ls==0.3.0
faker==8.14.0
aiobotocore==1.3.3
boto3==1.17.106
botocore==1.20.106
s3fs==2021.7.0
statsmodels==0.12.2
rich==10.3.0
baytune==0.4.0
catboost==0.25.1
lightgbm==3.2.1
fairlearn==0.7.0